requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
bcrypt==4.1.3
gunicorn==22.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
fpdf2==2.7.9
Pillow==10.4.0
WTForms==3.1.2
SQLAlchemy==2.0.36
scikit-learn==1.5.2
numpy==1.26.4
joblib==1.4.2
